Fix overflow when counting -F filter shortcuts (#2061)

+++ crates/cli/src/frontend/arguments/parsed_args.rs
@@ -122,7 +122,7 @@ pub(crate) struct ParsedArgs {
     pub(crate) include_from: Vec<OsString>,
     pub(crate) filters: Vec<OsString>,
     pub(crate) cvs_exclude: bool,
-    pub(crate) rsync_filter_shortcuts: u8,
+    pub(crate) rsync_filter_shortcuts: usize,
     pub(crate) files_from: Vec<OsString>,
     pub(crate) from0: bool,
     pub(crate) info: Vec<OsString>,
